Package names changed and changed delete booking according to the customer.

+++ src/main/kotlin/com/example/api/service/BookingService.kt
@@ -1,12 +1,12 @@
-package com.example.api.Service
+package com.example.api.service
 //import com.example.api.*
-import com.example.api.Controller.request.BookingRequest
-import com.example.api.Exception.EntityNotFoundException
-import com.example.api.Model.Booking
+import com.example.api.controller.request.BookingRequest
+import com.example.api.exception.EntityNotFoundException
+import com.example.api.model.Booking
 //import com.example.api.Model.BookingRequest
-import com.example.api.Repository.BookingRepository
+import com.example.api.repository.BookingRepository
 //import com.example.api.Repository.BookingRequestRepository
-import com.example.api.Repository.CustomerRepository
+import com.example.api.repository.CustomerRepository
 import org.springframework.stereotype.Service
 
 
@@ -43,9 +43,14 @@ import org.springframework.stereotype.Service
 
 
     fun deleteBooking(customerId: Int,id: Int) {
+        //book.id=id
         val customer= customerRepository.findById(customerId)
+        val booking=bookingRepository.findByCustomerCustomerIdAndId(customerId,id)
         if(customer.isEmpty){
             throw EntityNotFoundException("customer not found.")
+        }
+        if(booking.isEmpty) {
+            throw EntityNotFoundException("booking not found.")
         }
         bookingRepository.deleteById(id)
     }
